requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
Brotli==1.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.15
decorator==5.2.1
distro==1.9.0
duckduckgo_search==8.1.1
executing==2.2.0
faiss-cpu==1.11.0.post1
fastapi==0.116.1
ffmpy==0.6.0
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.7.0
google_search_results==2.4.2
gradio==5.37.0
gradio_client==1.10.4
greenlet==3.2.3
groovy==0.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.4
idna==3.10
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.8.1
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.69
langchain-openai==0.3.28
langchain-tavily==0.2.9
langchain-text-splitters==0.3.8
langgraph==0.5.3
langgraph-checkpoint==2.1.0
langgraph-checkpoint-sqlite==2.0.10
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.73
langsmith==0.4.6
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.6.3
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.3.1
openai==1.96.1
orjson==3.11.0
ormsgpack==1.10.0
packaging==25.0
pandas==2.3.1
parso==0.8.4
pillow==11.3.0
platformdirs==4.3.8
primp==0.15.0
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
pure_eval==0.2.3
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pypdf==5.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pywin32==311
PyYAML==6.0.2
pyzmq==27.0.0
regex==2024.11.6
reportlab==4.4.2
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
ruff==0.12.3
safehttpx==0.1.6
semantic-version==2.10.0
serpapi==0.1.5
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sqlite-vec==0.1.6
stack-data==0.6.3
starlette==0.47.1
tenacity==9.1.2
tiktoken==0.9.0
tomlkit==0.13.3
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
